# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MSLO Inventory API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /warehouses/{id}:
    get:
      summary: Get warehouse
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      summary: Update warehouse
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Deactivate warehouse
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /warehouses/{id}/locations:
    get:
      summary: List warehouse locations
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Create warehouse location
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '201': { description: Created }
  /warehouses/{id}/locations/{locationId}:
    put:
      summary: Update warehouse location
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: locationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Deactivate warehouse location
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: locationId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /users/{id}:
    get:
      summary: Get user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    put:
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Deactivate user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...